+++ CHANGES.md
@@ -18,6 +18,7 @@ Grammars:
 - enh(scala) add missing `do` and `then` keyword (#3323) [Nicolas Stucki][]
 - enh(scala) add missing `enum`, `export` and `given` keywords (#3328) [Nicolas Stucki][]
 - enh(scala) remove symbol syntax and fix quoted code syntax (#3324) [Nicolas Stucki][]
+- enh(scala) add Scala 3 `extension` soft keyword (#3326) [Nicolas Stucki][]
 
 [Austin Schick]: https://github.com/austin-schick
 [Josh Goebel]: https://github.com/joshgoebel